separar los datos de prueba del firebaseConfig.ts y cambiar ejemplos en data
