+++ App.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect } from "react";
 import { StatusBar } from "expo-status-bar";
 import { StyleSheet, Text, View } from "react-native";
-import example from "./backend/example.ts";
+import data from "./backend/data";
 
 export default function App() {
   useEffect(() => {
-    example();
+    data();
   }, []);
 
   return (
